+++ @xen-orchestra/lite/src/views/new-vm/NewVmView.vue
@@ -4,7 +4,7 @@
       {{ $t('new-vm.add') }}
     </UiHeadBar>
     <div class="card-container">
-      <form @submit.prevent="createVM()">
+      <form @submit.prevent="createVm()">
         <UiCard>
           <!-- TEMPLATE SECTION -->
           <UiTitle>{{ $t('template') }}</UiTitle>
@@ -281,7 +281,7 @@
                 <template v-if="vmState.existingVdis && vmState.existingVdis.length > 0">
                   <tr v-for="(vdi, index) in vmState.existingVdis" :key="index">
                     <td>
-                      <FormSelect v-model="vdi.SR" disabled>
+                      <FormSelect v-model="vdi.SR">
                         <option v-for="sr in filteredSrs" :key="sr.$ref" :value="sr.$ref">
                           {{ `${sr.name_label} -` }}
                           {{
@@ -394,9 +394,7 @@
               accent="brand"
               size="medium"
               :busy="isBusy"
-              :disabled="
-                vmState.new_vm_template === undefined || !hasInstallSettings || !hasVdis || hasInvalidSrVdi || isBusy
-              "
+              :disabled="vmState.new_vm_template === undefined || !hasInstallSettings || hasInvalidSrVdi || isBusy"
             >
               {{ $t('create') }}
             </UiButton>
@@ -413,7 +411,7 @@ import FormInputWrapper from '@/components/form/FormInputWrapper.vue'
 import FormSelect from '@/components/form/FormSelect.vue'
 
 // XenAPI Store imports
-import type { XenApiVdi, XenApiVm } from '@/libs/xen-api/xen-api.types'
+import type { XenApiSr, XenApiVbd, XenApiVdi, XenApiVm } from '@/libs/xen-api/xen-api.types'
 import { useHostStore } from '@/stores/xen-api/host.store.ts'
 import { useNetworkStore } from '@/stores/xen-api/network.store'
 import { usePifStore } from '@/stores/xen-api/pif.store'
@@ -460,9 +458,8 @@ import {
   faTrash,
 } from '@fortawesome/free-solid-svg-icons'
 
-import { type DOMAIN_TYPE, OPAQUE_REF, VBD_TYPE } from '@vates/types/common'
-
 // Vue imports
+import { OPAQUE_REF, VBD_TYPE, type DOMAIN_TYPE } from '@vates/types'
 import defer, { type Defer } from 'golike-defer'
 import { computed, reactive, ref } from 'vue'
 import { useI18n } from 'vue-i18n'
@@ -651,8 +648,6 @@ const hasInstallSettings = computed(() => {
   }
 })
 
-const hasVdis = computed(() => vmState.vdis.length > 0 || vmState.existingVdis.length > 0)
-
 const hasInvalidSrVdi = computed(() => vmState.vdis.some(vdi => vdi.SR && vdi.SR === OPAQUE_REF.EMPTY))
 
 const getDefaultNetworks = (template: XenApiVm) => {
@@ -673,11 +668,11 @@ const getDefaultNetworks = (template: XenApiVm) => {
 }
 
 const getExistingInterface = (template: XenApiVm) => {
-  const existingInterfaces = []
+  const interfaces = []
   const defaultNetwork = getDefaultNetworks(template)[0]
 
   if (template.VIFs.length === 0 && defaultNetwork) {
-    existingInterfaces.push({
+    interfaces.push({
       interface: defaultNetwork.$ref,
       macAddress: '',
     })
@@ -686,21 +681,26 @@ const getExistingInterface = (template: XenApiVm) => {
   template.VIFs.forEach(ref => {
     const vif = getVifByOpaqueRef(ref)
     if (vif) {
-      existingInterfaces.push({
+      interfaces.push({
         interface: getNetworkByOpaqueRef(vif.network)?.$ref || '',
         macAddress: vif.MAC || '',
       })
     }
   })
 
-  return existingInterfaces
+  return interfaces
 }
 
 const addNetworkInterface = () => {
-  if (!vmState.new_vm_template) return
+  if (!vmState.new_vm_template) {
+    return
+  }
 
   const defaultNetworks = getDefaultNetworks(vmState.new_vm_template)
-  if (defaultNetworks.length === 0) return
+
+  if (defaultNetworks.length === 0) {
+    return
+  }
 
   const defaultNetwork = defaultNetworks[0]
 
@@ -711,7 +711,7 @@ const addNetworkInterface = () => {
 }
 
 const getVdis = (template: XenApiVm) => {
-  const vdisArray = [] as Vdi[]
+  const vdis = [] as Vdi[]
 
   const parser = new DOMParser()
   const xmlString = template.other_config.disks
@@ -724,14 +724,14 @@ const getVdis = (template: XenApiVm) => {
     return []
   }
 
-  vdisArray.push({
+  vdis.push({
     name_label: (vmState.name || 'disk') + '_' + generateRandomString(4),
     name_description: 'Created by XO',
     size: bytesToGiB(Number(size)),
     SR: defaultSr.value,
   })
 
-  return vdisArray
+  return vdis
 }
 
 const getExistingVdis = (template: XenApiVm) => {
@@ -761,7 +761,10 @@ const getExistingVdis = (template: XenApiVm) => {
 
 const onTemplateChange = () => {
   const template = vmState.new_vm_template
-  if (!template) return
+
+  if (!template) {
+    return
+  }
 
   const {
     name_label,
@@ -819,7 +822,7 @@ const vmCreationParams = computed(() => ({
 const xapi = useXenApiStore().getXapi()
 
 // TODO move in job system
-const _createVm = async ($defer: Defer) => {
+const _createVm = defer(async ($defer: Defer) => {
   const templateRef = vmCreationParams.value.template!
   const newVmName = vmCreationParams.value.name_label
   const selectedVdiRef = vmCreationParams.value.installRepository
@@ -886,6 +889,7 @@ const _createVm = async ($defer: Defer) => {
     if (vm.domain_type === 'hvm') {
       if ((newVdis.length === 0 && existingVdis.length === 0) || installMethod === 'network') {
         const { order } = vm.HVM_boot_params
+
         await xapi.call('VM.set_HVM_boot_params', [vmRefs[0], { order: order ? 'n' + order.replace('n', '') : 'ncd' }])
       }
     } else {
@@ -905,6 +909,7 @@ const _createVm = async ($defer: Defer) => {
 
       for (const vbdRef of vm.VBDs) {
         const type = await xapi.getField<VBD_TYPE>('VBD', vbdRef, 'type')
+
         if (type === VBD_TYPE.CD) {
           await xapi.vbd.insert(vbdRef, selectedVdiRef)
           await xapi.vbd.setBootable(vbdRef, true)
@@ -943,7 +948,6 @@ const _createVm = async ($defer: Defer) => {
     }
 
     // EDIT VBDs IN CASE OF EXISTING VDIs
-    // TODO edit SR
     for (const vbdRef of vm.VBDs) {
       const vdiRef = await xapi.getField<XenApiVdi['$ref']>('VBD', vbdRef, 'VDI')
       const type = await xapi.getField<VBD_TYPE>('VBD', vbdRef, 'type')
@@ -953,19 +957,55 @@ const _createVm = async ($defer: Defer) => {
         hasBootableDisk = bootable
       }
 
-      if (!vbdRef || type === VBD_TYPE.CD) {
+      if (vdiRef === OPAQUE_REF.EMPTY || type === VBD_TYPE.CD) {
         continue
       }
 
       for (const vdi of existingVdis) {
+        const srRef = await xapi.getField<XenApiSr['$ref']>('VDI', vdiRef, 'SR')
+        const vbdRefs = await xapi.getField<XenApiVbd['$ref'][]>('VDI', vdiRef, 'VBDs')
+
         await xapi.vdi.setNameLabel(vdiRef, vdi.name_label)
         await xapi.vdi.setNameDescription(vdiRef, vdi.name_description)
+
+        // EDIT SR: Migrate VDI or create new one
+        if (vdi.SR !== srRef) {
+          try {
+            await xapi.vdi.poolMigrate(vdiRef, vdi.SR)
+          } catch (error) {
+            if (!(error instanceof Error)) {
+              throw error
+            }
+
+            const { message } = error
+
+            if (
+              message !== 'NO_HOSTS_AVAILABLE' &&
+              message !== 'LICENCE_RESTRICTION' &&
+              message !== 'VDI_NEEDS_VM_FOR_MIGRATE'
+            ) {
+              throw error
+            }
+
+            const newVdiRef = await xapi.vdi.copy(vdiRef, vdi.SR)
+
+            if (vbdRefs) {
+              for (const vbdRef of vbdRefs) {
+                const vbd = await xapi.getField<XenApiVbd>('VBD', vbdRef, 'record')
+
+                await xapi.vbd.delete(vbdRef)
+                await xapi.vbd.create({ ...vbd, vmRefs, vdiRef: newVdiRef[0] })
+              }
+            }
+
+            await xapi.vdi.delete(vdiRef)
+          }
+        }
       }
     }
 
     // VDIs AND VBDs CREATION
     let index = 0
-
     for (const vdi of newVdis) {
       const vdiRef = await xapi.vdi.create({
         ...vdi,
@@ -991,17 +1031,14 @@ const _createVm = async ($defer: Defer) => {
       $defer.onFailure(() => xapi.vbd.delete(vbdRef))
     }
 
-    // BOOT VM AFTER CREATION
-    if (vmCreationParams.value.bootAfterCreate) {
-      await xapi.vm.start(vmRefs)
-    }
-
     isBusy.value = false
 
     await router.push({
-      name: 'vm.console',
+      name: 'vm.default',
       params: { uuid: vm.uuid },
     })
+
+    return vmRefs
   } catch (error) {
     isBusy.value = false
 
@@ -1011,9 +1048,16 @@ const _createVm = async ($defer: Defer) => {
 
     throw error
   }
-}
+})
 
-const createVM = defer(_createVm)
+const createVm = async () => {
+  const vmRefs = await _createVm()
+
+  // BOOT VM AFTER CREATION
+  if (vmCreationParams.value.bootAfterCreate) {
+    await xapi.vm.start(vmRefs)
+  }
+}
 </script>
 
 <style scoped lang="postcss">