+++ @xen-orchestra/lite/src/views/vm/VmSystemView.vue
@@ -16,6 +16,13 @@
 </template>
 
 <script lang="ts" setup>
+import VmGeneralInfo from '@/components/vm/system/VmGeneralInfo.vue'
+import VmGraphics from '@/components/vm/system/VmGraphics.vue'
+import VmManagement from '@/components/vm/system/VmManagement.vue'
+import VmResource from '@/components/vm/system/VmResource.vue'
+import VmStorageConfig from '@/components/vm/system/VmStorageConfig.vue'
+import VmSystemNetworking from '@/components/vm/system/VmSystemNetworking.vue'
+import VmVirtualizationAndBoot from '@/components/vm/system/VmVirtualizationAndBoot.vue'
 import type { XenApiVm } from '@/libs/xen-api/xen-api.types'
 import { usePageTitleStore } from '@/stores/page-title.store'
 import { useVmStore } from '@/stores/xen-api/vm.store'
@@ -25,13 +32,6 @@ import VtsObjectNotFoundHero from '@core/components/state-hero/VtsObjectNotFound
 import { computed } from 'vue'
 import { useI18n } from 'vue-i18n'
 import { useRoute } from 'vue-router'
-import VmGeneralInfo from './system/VmGeneralInfo.vue'
-import VmGraphics from './system/VmGraphics.vue'
-import VmManagement from './system/VmManagement.vue'
-import VmResource from './system/VmResource.vue'
-import VmStorageConfig from './system/VmStorageConfig.vue'
-import VmSystemNetworking from './system/VmSystemNetworking.vue'
-import VmVirtualizationAndBoot from './system/VmVirtualizationAndBoot.vue'
 
 const route = useRoute()
 const { getByUuid } = useVmStore().subscribe()