+++ @xen-orchestra/web/src/pages/vm/[id]/system.vue
@@ -1,17 +1,17 @@
 <template>
-  <div class="system" :class="{ mobile: uiStore.isMobile }">
-    <div class="column">
+  <VtsColumns>
+    <VtsColumn>
       <VmGeneralInfo :vm />
       <VmNetworking :vm />
       <VmStorageConfig :vm />
       <VmResource :vm />
-    </div>
-    <div class="column">
+    </VtsColumn>
+    <VtsColumn>
       <VmVirtualizationAndBoot :vm />
       <VmManagement :vm />
       <VmGraphics :vm />
-    </div>
-  </div>
+    </VtsColumn>
+  </VtsColumns>
 </template>
 
 <script setup lang="ts">
@@ -23,33 +23,8 @@ import VmResource from '@/components/vm/system/VmResource.vue'
 import VmStorageConfig from '@/components/vm/system/VmStorageConfig.vue'
 import VmVirtualizationAndBoot from '@/components/vm/system/VmVirtualizationAndBoot.vue'
 import type { XoVm } from '@/types/xo/vm.type'
-import { useUiStore } from '@core/stores/ui.store'
+import VtsColumn from '@core/components/column/VtsColumn.vue'
+import VtsColumns from '@core/components/columns/VtsColumns.vue'
 
 defineProps<{ vm: XoVm }>()
-const uiStore = useUiStore()
 </script>
-
-<style scoped lang="postcss">
-.system {
-  display: flex;
-  gap: 0.8rem;
-  padding: 0.8rem;
-  width: 100%;
-  flex-direction: row;
-
-  .column {
-    display: flex;
-    flex-direction: column;
-    gap: 0.8rem;
-    width: 50%;
-  }
-
-  &.mobile {
-    flex-direction: column;
-
-    .column {
-      width: 100%;
-    }
-  }
-}
-</style>